Remove documentation and unused files from project repository
